Move test methods to utils module.
