+++ src/msmhelper/utils/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""  --8<-- "README.md" """
+
+from . import tests
